+++ docs/changelog.rst
@@ -1,4 +1,10 @@
-v6.1.0 (To be release)
+v6.2.0 (October)
+=========================
+Feat: Fixed several bugs of ZpaPortalTalker methods.(by `Sergio Pereira <mailto:spereira@zscaler.com>`_)
+Fix: Add Method to obtain ZPA services and druid cloud (by `Sergio Pereira <mailto:spereira@zscaler.com>`_)
+
+
+v6.1.0 (October)
 =========================
 Feat: Added cloud_connector API Talker for Cloud and Branch Connector (by `Dominic Schimanski <mailto:dschimanski@zscaler.com>)
 Feat: Added list_subcloud methods to ZiaPortalTalker (by `Patrick de Niet`)